Better performance for smaller number of threads

+++ src/IO/ReadBufferFromS3.h
@@ -59,7 +59,7 @@ private:
 
     ReadSettings read_settings;
     bool use_external_buffer;
-    size_t last_offset;
+    off_t last_offset;
 };
 
 }
